+++ tests/rest-apis/blockchains/BTC/btc.test.js
@@ -4,6 +4,7 @@ const PaymentForwarding = require('./payment-forwarding.test');
 const Transaction = require('./transaction.test');
 const Wallet = require('./wallet.test');
 const Webhook = require('./webhook.test');
+const OmniLayer = require('./omni-layer.test');
 
 async function BTC(caClient) {
     console.log('\n::BTC');
@@ -14,6 +15,7 @@ async function BTC(caClient) {
     await Transaction(caClient);
     await Wallet(caClient);
     await Webhook(caClient);
+    await OmniLayer(caClient);
 }
 
 module.exports = caClient => BTC(caClient);