+++ tests/rest-apis/blockchains/ETC/blockchain.test.js
@@ -2,7 +2,7 @@ async function Blockchain(caClient) {
     console.log('\n::ETC Blockchain');
 
     const chainInfo = await caClient.BC.ETC.blockchain.getInfo().then(response => response.payload);
-    const block = await caClient.BC.ETC.blockchain.getBlockByHeight(chainInfo.height - 120).then(response => response.payload);
+    const block = await caClient.BC.ETC.blockchain.getBlockByHeight(chainInfo.height - 1200).then(response => response.payload);
 
     await caClient.BC.ETC.blockchain.getBlockByHash(block.hash);
     await caClient.BC.ETC.blockchain.getLatestBlock();